chatbot/requirements.txt
# SPEAR Climate Chatbot Dependencies

# Core dependencies
streamlit==1.40.2
python-dotenv==1.0.1
Pillow>=9.0.0

# Data handling
xarray==2024.11.0
netCDF4==1.7.2
h5netcdf==1.4.0
s3fs==2024.12.0
boto3==1.35.86
fsspec==2024.12.0

# Plotting
matplotlib==3.10.0
numpy==2.2.1

# PDF processing
PyPDF2==3.0.1

# Async support
aiohttp==3.11.11
httpx==0.27.2

# MCP (Model Context Protocol)
# Match server version for compatibility
mcp==1.13.0

# LLM API clients
anthropic>=0.39.0
google-generativeai>=0.8.0

# Authentication (used when AUTH_ENABLED=true)
streamlit-authenticator>=0.3.0
PyYAML>=6.0
bcrypt>=4.0.0
